Clean up np.type deprecation warnings
Some np.type names have been deprecated since numpy 1.20.0
release, the Python builtin should be used instead. See https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations

+++ invisible_cities/core/core_functions_test.py
@@ -374,7 +374,7 @@ def test_binedges_from_bincenters_exceptions():
         core.binedges_from_bincenters(centers, range=range)
 
 
-@given(x = arrays(np.float, 100, elements=floats(min_value=-500, max_value=500)),
+@given(x = arrays(float, 100, elements=floats(min_value=-500, max_value=500)),
        value = floats(min_value=-500, max_value=500))
 def test_find_nearest(x, value):
